authenticator/requirements.txt
Flask==1.1.1
flask-mysql==1.4.0
pycryptodome==3.9.7
cryptography==2.8
